requirements.txt
streamlit
python-docx
PyMuPDF
google-generativeai
python-dotenv
scikit-learn
fuzzywuzzy[speedup]
